+++ backend/runner/types.go
@@ -1,0 +1,11 @@
+package runner
+
+type AppCategory struct {
+	Name string    `json:"name"`
+	Apps []AppItem `json:"apps"`
+}
+
+type AppItem struct {
+	Name string `json:"name"`
+	ID   string `json:"id"`
+}